\name{factoall}
\alias{factoall}
\title{factoall}
\description{Function to convert factors in a data frame to the appropriate format (character, numeric...)}
\usage{factoall(d, ncmax=10) }
\arguments{
\item{d}{Data frame to be converted}
\item{ncmax}{Maximum number of characters for numeric fields}
}
\value{required
}
\author{Eric Letouze}
\keyword{methods}
