\name{factochar}
\alias{factochar}
\title{factochar}
\description{Function to convert factors factors in a data frame to characters}
\usage{factochar(d) }
\arguments{
\item{d}{Data frame to be converted}
}
\value{Data frame with factors converted to characters
}
\author{Eric Letouze}
\keyword{methods}
