\name{geco.changeRange}
\alias{geco.changeRange}
\title{geco.changeRange}
\description{Function that linearly changes the range of a numeric vector}
\usage{geco.changeRange(v, newmin=1, newmax=10) }
\arguments{
\item{v}{Vector to be transformed}
\item{newmin}{New min value}
\item{newmax}{New max value}
}
\value{Vector linearly transformed with the new range
}
\author{Eric Letouze}
\keyword{methods}
