\name{geco.density}
\alias{geco.density}
\title{geco.density}
\description{wrapping of the function 'density' adding down and top points to the result }
\usage{geco.density(x,doplot=FALSE,pc=.05,...)}
\arguments{
\item{x}{a numeric vector}
\item{doplot}{boolean}
\item{pc}{...}
}
\value{a list with 2 elements : down and top points
}
\author{Eric Letouze}
\keyword{methods}
