\name{geco.groupMat}
\alias{geco.groupMat}
\title{geco.groupMat}
\description{This function calculates the average values in a matrix per groups of lines or columns}
\usage{geco.groupMat(mat=NA,						  margin=1,						  groups=NA,						  method="mean"						  )}
\arguments{
\item{mat}{Matrix to be averaged}
\item{margin}{Margin (1=lines, 2=columns)}
\item{groups}{List of groups}
\item{method}{mean or median}
}
\value{Averaged matrix
}
\author{Eric Letouze}
\keyword{methods}
