\name{geco.H1proportion}
\alias{geco.H1proportion}
\title{geco.H1proportion}
\description{Function that estimates the number of test under H1 hypothesis using Storey method}
\usage{geco.H1proportion(pv=NA,				              lambda=0.5				             ) }
\arguments{
\item{pv}{Vector of pvalues}
\item{lambda}{Lambda parameter for Storey method}
}
\value{Proportion of tests under H1 hypothesis
}
\author{Eric Letouze}
\keyword{methods}
