\name{geco.makeRd}
\alias{geco.makeRd}
\title{geco.makeRd}
\description{generate Rd files for a given parsed \code{R} file with annotations of functions following the makeRd format, as generated by \link[geco.utils]{geco.getMakeRd}}
\usage{geco.makeRd(file, directory=NULL, verbose=FALSE,... )}
\arguments{
\item{ file}{the given \file{.R} file to parse}
\item{directory}{the directory where the \file{.Rd} file is generated}
\item{verbose}{if \code{TRUE}, it lists the files created. \code{FALSE} by default}
\item{...}{non-matched further arguments are disregarded}
}
\value{generate a \file{.Rd} file for a given parsed file
}
\author{Eric Letouze}
\keyword{documentation}
